requirements.txt
agent-framework
feedparser
python-dotenv
httpx
pyyaml
pydantic
